refactor: extract view, util, create/del actions
actions file was getting pretty big
