+++ web/src/modules/topic/store/actions.ts
@@ -1,224 +1,7 @@
-import { emitter } from "../../../common/event";
-import { getClaimDiagramId, getImplicitLabel, parseClaimDiagramId } from "../utils/claim";
-import {
-  ArguableType,
-  Node,
-  RelationDirection,
-  Score,
-  buildEdge,
-  buildNode,
-  findArguable,
-  findNode,
-  getNodesComposedBy,
-  layoutVisibleComponents,
-} from "../utils/diagram";
-import { Relation, canCreateEdge, getConnectingEdge, getRelation } from "../utils/edge";
-import { NodeType, children, edges } from "../utils/node";
-import { TopicStoreState, initialState, problemDiagramId, useTopicStore } from "./store";
-
-export const getState = () => {
-  return useTopicStore.getState();
-};
-
-export const setState = (state: TopicStoreState) => {
-  useTopicStore.setState(() => state);
-};
-
-export const resetState = () => {
-  useTopicStore.setState(initialState, false, "resetState");
-};
-
-export const undo = () => {
-  useTopicStore.temporal.getState().undo();
-};
-
-export const redo = () => {
-  useTopicStore.temporal.getState().redo();
-};
-
-const getActiveDiagram = (state: TopicStoreState) => {
-  const activeDiagramId = state.activeClaimDiagramId ?? problemDiagramId;
-  return state.diagrams[activeDiagramId];
-};
-
-const createNode = (state: TopicStoreState, toNodeType: NodeType) => {
-  /* eslint-disable functional/immutable-data, no-param-reassign */
-  const newNodeId = `${state.nextNodeId++}`;
-  /* eslint-enable functional/immutable-data, no-param-reassign */
-
-  const activeDiagram = getActiveDiagram(state);
-  const newNode = buildNode({ id: newNodeId, type: toNodeType, diagramId: activeDiagram.id });
-
-  /* eslint-disable functional/immutable-data, no-param-reassign */
-  activeDiagram.nodes.push(newNode);
-  /* eslint-enable functional/immutable-data, no-param-reassign */
-
-  return newNode;
-};
-
-// if adding a criterion, connect to solutions
-// if adding a solution, connect to criteria
-const connectCriteriaToSolutions = (state: TopicStoreState, newNode: Node, problemNode: Node) => {
-  const problemDiagram = state.diagrams[problemDiagramId]; // solutions & criteria only will be in the problem diagram
-
-  const targetRelation: Relation =
-    newNode.type === "criterion"
-      ? { child: "solution", name: "solves", parent: "problem" }
-      : { child: "criterion", name: "criterion for", parent: "problem" };
-
-  const newCriterionEdges = problemDiagram.edges
-    .filter(
-      (edge) =>
-        edge.source === problemNode.id &&
-        edge.label === targetRelation.name &&
-        findNode(problemDiagram, edge.target).type === targetRelation.child
-    )
-    .map((edge) => {
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      const newCriterionEdgeId = `${state.nextEdgeId++}`;
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-
-      const sourceNodeId = newNode.type === "criterion" ? newNode.id : edge.target;
-      const targetNodeId = newNode.type === "criterion" ? edge.target : newNode.id;
-
-      return buildEdge(
-        newCriterionEdgeId,
-        sourceNodeId,
-        targetNodeId,
-        "embodies",
-        problemDiagramId
-      );
-    });
-
-  /* eslint-disable functional/immutable-data, no-param-reassign */
-  problemDiagram.edges.push(...newCriterionEdges);
-  /* eslint-enable functional/immutable-data, no-param-reassign */
-};
-
-interface AddNodeProps {
-  fromNodeId: string;
-  as: RelationDirection;
-  toNodeType: NodeType;
-  relation: Relation;
-}
-
-export const addNode = ({ fromNodeId, as, toNodeType, relation }: AddNodeProps) => {
-  useTopicStore.setState(
-    (state) => {
-      const activeDiagram = getActiveDiagram(state);
-      const fromNode = findNode(activeDiagram, fromNodeId);
-
-      // create and connect node
-      const newNode = createNode(state, toNodeType);
-
-      const parentNode = as === "parent" ? newNode : fromNode;
-      const childNode = as === "parent" ? fromNode : newNode;
-      createEdgeAndImpliedEdges(state, parentNode, childNode, relation);
-
-      // connect criteria
-      if (["criterion", "solution"].includes(newNode.type) && fromNode.type === "problem") {
-        connectCriteriaToSolutions(state, newNode, fromNode);
-      }
-
-      // re-layout
-      const layoutedDiagram = layoutVisibleComponents(activeDiagram);
-
-      // trigger event so viewport can be updated.
-      // seems like there should be a cleaner way to do this - perhaps custom zustand middleware to emit for any action
-      emitter.emit("addNode", findNode(layoutedDiagram, newNode.id));
-
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      activeDiagram.nodes = layoutedDiagram.nodes;
-      activeDiagram.edges = layoutedDiagram.edges;
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-    },
-    false,
-    "addNode" // little gross, seems like this should be inferrable from method name
-  );
-};
-
-const createEdgesImpliedByComposition = (
-  state: TopicStoreState,
-  parent: Node,
-  child: Node,
-  relation: Relation
-) => {
-  const diagram = state.diagrams[parent.data.diagramId];
-
-  const nodesComposedByParent = getNodesComposedBy(parent, diagram);
-  nodesComposedByParent.forEach((composedNode) => {
-    const relationForComposed = getRelation(composedNode.type, relation.child, relation.name);
-    if (!relationForComposed) return;
-
-    createEdgeAndImpliedEdges(state, composedNode, child, relationForComposed);
-  });
-
-  const nodesComposedByChild = getNodesComposedBy(child, diagram);
-  nodesComposedByChild.forEach((composedNode) => {
-    const relationForComposed = getRelation(relation.parent, composedNode.type, relation.name);
-    if (!relationForComposed) return;
-
-    createEdgeAndImpliedEdges(state, parent, composedNode, relationForComposed);
-  });
-};
-
-// see algorithm pseudocode & example at https://github.com/amelioro/ameliorate/issues/66#issuecomment-1465078133
-const createEdgeAndImpliedEdges = (
-  state: TopicStoreState,
-  parent: Node,
-  child: Node,
-  relation: Relation
-) => {
-  const diagram = state.diagrams[parent.data.diagramId];
-
-  // assumes only one edge can exist between two notes - future may allow multiple edges of different relation type
-  if (getConnectingEdge(parent, child, diagram.edges) !== undefined) return diagram.edges;
-
-  /* eslint-disable functional/immutable-data, no-param-reassign */
-  const newEdgeId = `${state.nextEdgeId++}`;
-  /* eslint-enable functional/immutable-data, no-param-reassign */
-
-  const newEdge = buildEdge(newEdgeId, parent.id, child.id, relation.name, diagram.id);
-
-  /* eslint-disable functional/immutable-data, no-param-reassign */
-  diagram.edges = diagram.edges.concat(newEdge);
-  /* eslint-enable functional/immutable-data, no-param-reassign */
-
-  // indirectly recurses by calling this method after determining which implied edges to create
-  // note: modifies diagram.edges through `state` (via the line above)
-  createEdgesImpliedByComposition(state, parent, child, relation);
-
-  return diagram.edges;
-};
-
-export const connectNodes = (parentId: string | null, childId: string | null) => {
-  useTopicStore.setState(
-    (state) => {
-      const activeDiagram = getActiveDiagram(state);
-
-      const parent = activeDiagram.nodes.find((node) => node.id === parentId);
-      const child = activeDiagram.nodes.find((node) => node.id === childId);
-      if (!parent || !child) throw new Error("parent or child not found");
-
-      if (!canCreateEdge(activeDiagram, parent, child)) return;
-
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion -- canCreateEdge ensures relation is valid
-      const relation = getRelation(parent.type, child.type)!;
-
-      // modifies diagram.edges through `state`
-      createEdgeAndImpliedEdges(state, parent, child, relation);
-
-      const layoutedDiagram = layoutVisibleComponents(activeDiagram);
-
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      activeDiagram.nodes = layoutedDiagram.nodes;
-      activeDiagram.edges = layoutedDiagram.edges;
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-    },
-    false,
-    "connectNodes"
-  );
-};
+import { getClaimDiagramId, parseClaimDiagramId } from "../utils/claim";
+import { ArguableType, Score, findArguable, findNode } from "../utils/diagram";
+import { problemDiagramId, useTopicStore } from "./store";
+import { getActiveDiagram } from "./utils";
 
 export const deselectNodes = () => {
   useTopicStore.setState(
@@ -291,80 +74,6 @@ export const doesDiagramExist = (diagramId: string) => {
   return Object.keys(useTopicStore.getState().diagrams).includes(diagramId);
 };
 
-export const viewOrCreateClaimDiagram = (arguableId: string, arguableType: ArguableType) => {
-  useTopicStore.setState(
-    (state) => {
-      const diagramId = getClaimDiagramId(arguableId, arguableType);
-
-      // create claim diagram if it doesn't exist
-      if (!doesDiagramExist(diagramId)) {
-        const activeDiagram = getActiveDiagram(state);
-        const arguable = findArguable(activeDiagram, arguableId, arguableType);
-        const label = getImplicitLabel(arguableId, arguableType, activeDiagram);
-
-        /* eslint-disable functional/immutable-data, no-param-reassign */
-        const newNode = buildNode({
-          id: `${state.nextNodeId++}`,
-          label: label,
-          score: arguable.data.score,
-          type: "rootClaim",
-          diagramId: diagramId,
-        });
-
-        state.diagrams[diagramId] = {
-          id: diagramId,
-          nodes: [newNode],
-          edges: [],
-          type: "claim",
-        };
-        /* eslint-enable functional/immutable-data, no-param-reassign */
-      }
-
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      state.activeClaimDiagramId = diagramId;
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-    },
-    false,
-    "viewOrCreateClaimDiagram"
-  );
-};
-
-export const viewClaimDiagram = (diagramId: string) => {
-  useTopicStore.setState(
-    (state) => {
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      state.activeClaimDiagramId = diagramId;
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-    },
-    false,
-    "viewClaimDiagram"
-  );
-};
-
-export const closeClaimDiagram = () => {
-  useTopicStore.setState(
-    (state) => {
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      state.activeClaimDiagramId = null;
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-    },
-    false,
-    "closeClaimDiagram"
-  );
-};
-
-export const closeTable = () => {
-  useTopicStore.setState(
-    (state) => {
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      state.activeTableProblemId = null;
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-    },
-    false,
-    "closeTable"
-  );
-};
-
 export const setNodeLabel = (nodeId: string, value: string) => {
   useTopicStore.setState(
     (state) => {
@@ -377,116 +86,5 @@ export const setNodeLabel = (nodeId: string, value: string) => {
     },
     false,
     "setNodeLabel"
-  );
-};
-
-export const toggleShowCriteria = (problemNodeId: string, show: boolean) => {
-  useTopicStore.setState(
-    (state) => {
-      const problemDiagram = state.diagrams[problemDiagramId]; // criteria nodes only live in problem diagram
-
-      const node = findNode(problemDiagram, problemNodeId);
-      if (node.type !== "problem") throw new Error("node is not a problem");
-
-      const criteria = children(node, problemDiagram).filter((child) => child.type === "criterion");
-
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      criteria.forEach((criterion) => (criterion.data.showing = show));
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-
-      const layoutedDiagram = layoutVisibleComponents(problemDiagram); // depends on showCriteria having been updated
-
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      problemDiagram.nodes = layoutedDiagram.nodes;
-      problemDiagram.edges = layoutedDiagram.edges;
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-    },
-    false,
-    "toggleShowCriteria"
-  );
-};
-
-export const deleteNode = (nodeId: string) => {
-  useTopicStore.setState(
-    (state) => {
-      const activeDiagram = getActiveDiagram(state);
-
-      const node = findNode(activeDiagram, nodeId);
-
-      if (node.type === "rootClaim") {
-        /* eslint-disable functional/immutable-data, no-param-reassign */
-        // eslint-disable-next-line @typescript-eslint/no-dynamic-delete -- consider using a map instead of an object?
-        delete state.diagrams[activeDiagram.id];
-        state.activeClaimDiagramId = null;
-        /* eslint-enable functional/immutable-data, no-param-reassign */
-        return;
-      }
-
-      const nodeEdges = edges(node, activeDiagram);
-      const childDiagramId = getClaimDiagramId(node.id, "node");
-
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      activeDiagram.nodes = activeDiagram.nodes.filter((node) => node.id !== nodeId);
-      activeDiagram.edges = activeDiagram.edges.filter((edge) => !nodeEdges.includes(edge));
-      // eslint-disable-next-line @typescript-eslint/no-dynamic-delete -- consider using a map instead of an object?
-      delete state.diagrams[childDiagramId];
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-
-      const layoutedDiagram = layoutVisibleComponents(activeDiagram);
-
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      activeDiagram.nodes = layoutedDiagram.nodes;
-      activeDiagram.edges = layoutedDiagram.edges;
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-    },
-    false,
-    "deleteNode"
-  );
-};
-
-export const deleteEdge = (edgeId: string) => {
-  useTopicStore.setState(
-    (state) => {
-      const activeDiagram = getActiveDiagram(state);
-
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      activeDiagram.edges = activeDiagram.edges.filter((edge) => edge.id !== edgeId);
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-
-      const layoutedDiagram = layoutVisibleComponents(activeDiagram);
-
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      activeDiagram.nodes = layoutedDiagram.nodes;
-      activeDiagram.edges = layoutedDiagram.edges;
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-    },
-    false,
-    "deleteEdge"
-  );
-};
-
-export const viewProblemDiagram = () => {
-  useTopicStore.setState(
-    (state) => {
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      state.activeTableProblemId = null;
-      state.activeClaimDiagramId = null;
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-    },
-    false,
-    "viewProblemDiagram"
-  );
-};
-
-export const viewCriteriaTable = (problemNodeId: string) => {
-  useTopicStore.setState(
-    (state) => {
-      /* eslint-disable functional/immutable-data, no-param-reassign */
-      state.activeTableProblemId = problemNodeId;
-      state.activeClaimDiagramId = null;
-      /* eslint-enable functional/immutable-data, no-param-reassign */
-    },
-    false,
-    "viewCriteriaTable"
   );
 };
